requirements.txt
falcon
gunicorn
yubico
pycrypto
